(refactor) dedupe dex num/usd/pct formatters into shared module
One shared components/dex/format.ts replaces the three divergent private
copies in PoolBrowser, PoolStats and lp-position. PoolStats' 24h-change and
APR cells now use pct() with the K% cap, matching the table, and a genuine
$0 renders as $0.00 instead of an em-dash (dash is only for missing data).
